+++ app/src/js/actions/index.js
@@ -955,3 +955,4 @@ export const filterReconciliationReport = (param) => ({ type: types.FILTER_RECON
 export const clearReconciliationReportFilter = (paramKey) => ({ type: types.CLEAR_RECONCILIATION_FILTER, paramKey });
 
 export const toggleSidebar = () => ({ type: types.TOGGLE_SIDEBAR });
+export const sortPersist = (tableId, sortBy) => ({ type: types.SORTS, tableId, sortBy });